*Delete section api bugfix

+++ KLS_API/KLS_API/Controllers/Travels/TravelsController.cs
@@ -109,7 +109,7 @@ namespace KLS_API.Controllers.Travels
         {
             try
             {
-                Travel travel = await _dbContext.Travels.Include(x => x.Sections).ThenInclude(y => y.Services).Include(x => x.Sections).ThenInclude(x => x.Clients).Include(x => x.Sections).ThenInclude(x => x.Substatus).ThenInclude(x=> x.Status).SingleOrDefaultAsync(x => x.Id == TravelId);
+                Travel travel = await _dbContext.Travels.Include(x => x.Sections).ThenInclude(y => y.Services).Include(x => x.Sections).ThenInclude(x => x.Clients).Include(x => x.Sections).ThenInclude(x => x.Substatus).ThenInclude(x => x.Status).SingleOrDefaultAsync(x => x.Id == TravelId);
 
                 if (travel is null)
                 {
@@ -341,7 +341,10 @@ namespace KLS_API.Controllers.Travels
                     return NotFound();
                 }
                 Oferta oferta = await _dbContext.Oferta.SingleOrDefaultAsync(x => x.SectionId == SectionId);
-                _dbContext.Oferta.Remove(oferta);
+                if (oferta != null)
+                {
+                    _dbContext.Oferta.Remove(oferta);
+                }
                 _dbContext.Sections.Remove(section);
                 await _dbContext.SaveChangesAsync();
                 return Ok(section);
